Renamed test.js to media.js
